+++ lib/bmg/operator/join.rb
@@ -42,6 +42,25 @@ module Bmg
         [ :join, left.to_ast, right.to_ast, on ]
       end
 
+    protected ### optimization
+
+      def _autowrap(type, options)
+        u_left,  left_replaced  = _unautowrap(left, options)
+        u_right, right_replaced = _unautowrap(right, options)
+        if (!left_replaced && !right_replaced)
+          super
+        else
+          u_left.join(u_right, on).autowrap(options)
+        end
+      end
+
+      def _unautowrap(operand, options)
+        return [operand, false] unless operand.is_a?(Operator::Autowrap)
+        return [operand, false] unless operand.same_options?(options)
+        [operand.send(:operand), true]
+      end
+      private :_unautowrap
+
     protected ### inspect
 
       def args